+++ src/hooks/api/useSaleDetailQuery.ts
